# contracts/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ActorContext": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "maxLength": 36,
            "minLength": 1,
            "title": "Actor Id",
            "type": "string"
          },
          "locale": {
            "enum": [
              "zh-CN",
              "en-US"
            ],
            "title": "Locale",
            "type": "string"
          },
          "role": {
            "enum": [
              "ADMIN",
              "ENGINEER"
            ],
            "title": "Role",
            "type": "string"
          },
          "theme": {
            "enum": [
              "light",
              "dark"
            ],
            "title": "Theme",
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "role",
          "locale",
          "theme"
        ],
        "title": "ActorContext",
        "type": "object"
      },
      "AuthSession": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "minLength": 16,
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "maximum": 86400,
            "minimum": 1,
            "title": "Expires In",
            "type": "integer"
          },
          "locale": {
            "enum": [
              "zh-CN",
              "en-US"
            ],
            "title": "Locale",
            "type": "string"
          },
          "refresh_token": {
            "minLength": 16,
            "title": "Refresh Token",
            "type": "string"
          },
          "theme": {
            "enum": [
              "light",
              "dark"
            ],
            "title": "Theme",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user",
          "locale",
          "theme"
        ],
        "title": "AuthSession",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "field_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field Path"
          },
          "message_key": {
            "title": "Message Key",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message_key",
          "request_id"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "RefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "minLength": 16,
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
...
      "User": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "id": {
            "maxLength": 36,
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "role": {
            "enum": [
              "ADMIN",
              "ENGINEER"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "is_active"
        ],
        "title": "User",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "Opaque access token issued by POST /auth/login. Token hash is stored server-side in auth_sessions and can be revoked immediately.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "PCB-CDSO API",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Authenticate with email/password; returns access token, refresh session, user, locale and theme. Failure must not distinguish unknown user from wrong password (both UNAUTHENTICATED).",
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            },
            "description": "Authenticated"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "value": {
                      "code": "UNAUTHENTICATED",
                      "details": {},
                      "field_path": null,
                      "message_key": "example.unauthenticated",
                      "request_id": "req_example_0123456789abcdef"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Invalid credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Revoke the current session.",
        "operationId": "auth_logout",
        "responses": {
          "204": {
            "description": "Session revoked"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "value": {
                      "code": "UNAUTHENTICATED",
                      "details": {},
                      "field_path": null,
                      "message_key": "example.unauthenticated",
                      "request_id": "req_example_0123456789abcdef"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid token."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Return the ActorContext injected from the Authorization header.",
        "operationId": "auth_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorContext"
                }
              }
            },
            "description": "Current actor"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "value": {
                      "code": "UNAUTHENTICATED",
                      "details": {},
                      "field_path": null,
                      "message_key": "example.unauthenticated",
                      "request_id": "req_example_0123456789abcdef"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing or invalid token."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Current actor",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Rotate access and refresh tokens. The previous tokens are invalidated immediately in the same transaction.",
        "operationId": "auth_refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSession"
                }
              }
            },
            "description": "Rotated session"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "UNAUTHENTICATED": {
                    "value": {
                      "code": "UNAUTHENTICATED",
                      "details": {},
                      "field_path": null,
                      "message_key": "example.unauthenticated",
                      "request_id": "req_example_0123456789abcdef"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Refresh token expired, revoked or user deactivated."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}